shim: add shim for phoneserver
    pthread's pthread_setschedparam() and pthread_getschedparam()
expected the first param which is pthread_t to be passed as value. But
some how the phoneserver binary passed pthread_t as a pointer. The old
behavior of bionic pthread is just ignore that invalid pthread_t and
return null. New behavior of bionic pthread is changed in O and cause
phoneserver to crash.

    This shim treads the pthread_t passed to pthread_setschedparam()
and pthread_getschedparam() as a pointer an passed the deferenced pointer
back to the original function.

Change-Id: I8d792675f1d78229e633e82e2f9e802f96f8639d